src/test/java/com/cafe/facade/waiter/WaiterFacadeImplTest.java
package com.cafe.facade.waiter;

import com.cafe.entity.*;
import com.cafe.facade.dto.CreateOrderRequestDto;
import com.cafe.facade.dto.CreateProductInOrderRequestDto;
import com.cafe.facade.dto.EditeOrderRequestDto;
import com.cafe.facade.dto.EditeProductInOrderRequestDto;
import com.cafe.facade.dto.entity.OrderDto;
import com.cafe.facade.dto.entity.ProductInOrderDto;
import com.cafe.facade.dto.entity.TableDto;
import com.cafe.facade.mapper.order.OrderMapper;
import com.cafe.facade.mapper.productinorder.ProductInOrderMapper;
import com.cafe.facade.mapper.table.TableMapper;
import com.cafe.service.core.order.CreateOrderParams;
import com.cafe.service.core.order.OrderService;
import com.cafe.service.core.order.UpdateOrderParams;
import com.cafe.service.core.product.ProductService;
import com.cafe.service.core.product.UpdateProductParams;
import com.cafe.service.core.productinorder.CreateProductInOrderParams;
import com.cafe.service.core.productinorder.ProductInOrderService;
import com.cafe.service.core.productinorder.UpdateProductInOrderParams;
import com.cafe.service.core.table.TableService;
import com.cafe.service.core.user.UserService;
import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import java.util.List;
import java.util.Optional;

import static org.mockito.Mockito.*;

@ExtendWith(MockitoExtension.class)
class WaiterFacadeImplTest {

    private WaiterFacade testSubject;

    @Mock
    private TableService tableService;
    @Mock
    private OrderService orderService;
    @Mock
    private ProductInOrderService productInOrderService;
    @Mock
    private ProductService productService;
    @Mock
    private UserService userService;
    @Mock
    private TableMapper tableMapper;
    @Mock
    private OrderMapper orderMapper;
    @Mock
    private ProductInOrderMapper productInOrderMapper;

    @BeforeEach
    public void init() {
        testSubject = new WaiterFacadeImpl(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testGetAssignedTablesWhenUsernameIsNull() {
        Assertions.assertThatThrownBy(() -> testSubject.getAssignedTables(null))
                .isExactlyInstanceOf(IllegalArgumentException.class);
    }

    @Test
    public void testCreateOrderWhenDtoIsNull() {
        Assertions.assertThatThrownBy(() -> testSubject.createOrder(null))
                .isExactlyInstanceOf(IllegalArgumentException.class);
    }

    @Test
    public void testCreateProductInOrderWhenDtoIsNull() {
        Assertions.assertThatThrownBy(() -> testSubject.createProductInOrder(null))
                .isExactlyInstanceOf(IllegalArgumentException.class);
    }

    @Test
    public void testEditProductInOrderWhenDtoIsNull() {
        Assertions.assertThatThrownBy(() -> testSubject.editProductInOrder(null))
                .isExactlyInstanceOf(IllegalArgumentException.class);
    }

    @Test
    public void testEditOrderWhenDtoIsNull() {
        Assertions.assertThatThrownBy(() -> testSubject.editeOrder(null))
                .isExactlyInstanceOf(IllegalArgumentException.class);
    }

    @Test
    public void testGetAssignedTablesWhenUserNotFound() {
        when(userService.findByUsername("username"))
                .thenReturn(Optional.empty());

        Assertions.assertThat(testSubject.getAssignedTables("username"))
                .isEqualTo(List.of(new TableDto(List.of("Not found user with username - username"))));

        verify(userService).findByUsername("username");
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testGetAssignedTables() {
        final User user = new User();
        user.setId(1L);
        when(userService.findByUsername("username"))
                .thenReturn(Optional.of(user));

        when(tableService.findByWaiterId(1L))
                .thenReturn(List.of(new Table()));

        when(tableMapper.map(new Table()))
                .thenReturn(new TableDto(List.of("test")));

        Assertions.assertThat(testSubject.getAssignedTables("username"))
                .isEqualTo(List.of(new TableDto(List.of("test"))));

        verify(userService).findByUsername("username");
        verify(tableService).findByWaiterId(1L);
        verify(tableMapper).map(new Table());
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testCreateOrderWhenTableNotFound() {
        when(tableService.findById(1L))
                .thenReturn(Optional.empty());

        Assertions.assertThat(
                testSubject.createOrder(new CreateOrderRequestDto(
                        1L
                ))
        ).isEqualTo(new OrderDto(List.of("Not found table with id - 1")));

        verify(tableService).findById(1L);
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testCreateOrderWhenTableHasOpenOrder() {
        when(tableService.findById(1L))
                .thenReturn(Optional.of(new Table()));

        final Order order = new Order();
        order.setOrderStatus(OrderStatus.OPEN);
        when(orderService.findByTableId(1L))
                .thenReturn(List.of(order));

        Assertions.assertThat(
                testSubject.createOrder(new CreateOrderRequestDto(
                        1L
                ))
        ).isEqualTo(new OrderDto(List.of("Table with id - 1 already has open order" )));

        verify(tableService).findById(1L);
        verify(orderService).findByTableId(1L);
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testCreateOrder() {
        when(tableService.findById(1L))
                .thenReturn(Optional.of(new Table()));

        final Order order = new Order();
        order.setOrderStatus(OrderStatus.CLOSED);
        when(orderService.findByTableId(1L))
                .thenReturn(List.of(order));

        when(orderService.create(new CreateOrderParams(1L, OrderStatus.OPEN)))
                .thenReturn(new Order());

        when(orderMapper.map(new Order()))
                .thenReturn(new OrderDto(List.of("test")));

        Assertions.assertThat(
                testSubject.createOrder(new CreateOrderRequestDto(
                        1L
                ))
        ).isEqualTo(new OrderDto(List.of("test")));

        verify(tableService).findById(1L);
        verify(orderService).findByTableId(1L);
        verify(orderService).create(
                new CreateOrderParams(1L, OrderStatus.OPEN)
        );
        verify(orderMapper).map(new Order());
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testCreateProductInOrderWhenOrderNotFound() {
        when(orderService.findById(1L))
                .thenReturn(Optional.empty());

        Assertions.assertThat(
                testSubject.createProductInOrder(new CreateProductInOrderRequestDto(
                        1L,
                        "product",
                        10L
                ))
        ).isEqualTo(new ProductInOrderDto(List.of("Not found order with id - 1")));

        verify(orderService).findById(1L);
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testCreateProductInOrderWhenProductNotFound() {
        when(orderService.findById(1L))
                .thenReturn(Optional.of(new Order()));

        when(productService.findByName("product"))
                .thenReturn(Optional.empty());

        Assertions.assertThat(
                testSubject.createProductInOrder(new CreateProductInOrderRequestDto(
                        1L,
                        "product",
                        10L
                ))
        ).isEqualTo(new ProductInOrderDto(List.of("Not found product with name - product")));

        verify(orderService).findById(1L);
        verify(productService).findByName("product");
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testCreateProductInOrderWhenNotEnoughProduct() {
        when(orderService.findById(1L))
                .thenReturn(Optional.of(new Order()));

        final Product product = new Product();
        product.setAmount(5L);
        when(productService.findByName("product"))
                .thenReturn(Optional.of(product));

        Assertions.assertThat(
                testSubject.createProductInOrder(new CreateProductInOrderRequestDto(
                        1L,
                        "product",
                        10L
                ))
        ).isEqualTo(new ProductInOrderDto(List.of("Not enough product - product in the amount of 10")));

        verify(orderService).findById(1L);
        verify(productService).findByName("product");
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testCreateProductInOrder() {
        when(orderService.findById(1L))
                .thenReturn(Optional.of(new Order()));

        final Product product = new Product();
        product.setId(1L);
        product.setName("product");
        product.setAmount(15L);
        when(productService.findByName("product"))
                .thenReturn(Optional.of(product));

        when(productInOrderService.create(new CreateProductInOrderParams(
                1L,
                1L,
                10L,
                ProductInOrderStatus.ACTIVE
        ))).thenReturn(new ProductInOrder());

        when(productInOrderMapper.map(new ProductInOrder()))
                .thenReturn(new ProductInOrderDto(List.of("test")));

        Assertions.assertThat(
                testSubject.createProductInOrder(new CreateProductInOrderRequestDto(
                        1L,
                        "product",
                        10L
                ))
        ).isEqualTo(new ProductInOrderDto(List.of("test")));

        verify(orderService).findById(1L);
        verify(productService).findByName("product");
        verify(productInOrderService).create(new CreateProductInOrderParams(
                1L,
                1L,
                10L,
                ProductInOrderStatus.ACTIVE
        ));
        verify(productService).update(new UpdateProductParams(
                1L,
                "product",
                5L
        ));
        verify(productInOrderMapper).map(new ProductInOrder());
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testEditProductInOrderWhenNotFound() {
        when(productInOrderService.findById(1L))
                .thenReturn(Optional.empty());

        Assertions.assertThat(
                testSubject.editProductInOrder(new EditeProductInOrderRequestDto(
                        1L,
                        "product",
                        10L,
                        ProductInOrderStatus.ACTIVE
                ))
        ).isEqualTo(new ProductInOrderDto(List.of("Not found product in order with id - 1")));

        verify(productInOrderService).findById(1L);
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testEditProductInOrderWhenAlreadyCancelled() {
        final ProductInOrder productInOrder = new ProductInOrder();
        productInOrder.setProductInOrderStatus(ProductInOrderStatus.CANCELLED);
        when(productInOrderService.findById(1L))
                .thenReturn(Optional.of(productInOrder));


        Assertions.assertThat(
                testSubject.editProductInOrder(new EditeProductInOrderRequestDto(
                        1L,
                        "product",
                        10L,
                        ProductInOrderStatus.ACTIVE
                ))
        ).isEqualTo(new ProductInOrderDto(List.of("Product in order with id - 1 already cancelled")));

        verify(productInOrderService).findById(1L);
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testEditProductInOrderWhenProductNotFound() {
        final ProductInOrder productInOrder = new ProductInOrder();
        productInOrder.setProductInOrderStatus(ProductInOrderStatus.ACTIVE);
        when(productInOrderService.findById(1L))
                .thenReturn(Optional.of(productInOrder));

        when(productService.findByName("product"))
                .thenReturn(Optional.empty());

        Assertions.assertThat(
                testSubject.editProductInOrder(new EditeProductInOrderRequestDto(
                        1L,
                        "product",
                        10L,
                        ProductInOrderStatus.ACTIVE
                ))
        ).isEqualTo(new ProductInOrderDto(List.of("Not found product with name - product")));

        verify(productInOrderService).findById(1L);
        verify(productService).findByName("product");
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testEditProductInOrderWhenSameProductNotEnough() {
        final Product product = new Product();
        product.setName("product");
        product.setAmount(9L);

        final ProductInOrder productInOrder = new ProductInOrder();
        productInOrder.setProduct(product);
        productInOrder.setAmount(5L);
        productInOrder.setProductInOrderStatus(ProductInOrderStatus.ACTIVE);
        when(productInOrderService.findById(1L))
                .thenReturn(Optional.of(productInOrder));


        when(productService.findByName("product"))
                .thenReturn(Optional.of(product));

        Assertions.assertThat(
                testSubject.editProductInOrder(new EditeProductInOrderRequestDto(
                        1L,
                        "product",
                        15L,
                        ProductInOrderStatus.ACTIVE
                ))
        ).isEqualTo(new ProductInOrderDto(List.of("Not enough product - product in the amount of 15")));

        verify(productInOrderService).findById(1L);
        verify(productService).findByName("product");
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testEditProductInOrderWhenOtherProductNotEnough() {
        final ProductInOrder productInOrder = new ProductInOrder();
        productInOrder.setProduct(new Product());
        productInOrder.setAmount(5L);
        productInOrder.setProductInOrderStatus(ProductInOrderStatus.ACTIVE);
        when(productInOrderService.findById(1L))
                .thenReturn(Optional.of(productInOrder));

        final Product product = new Product();
        product.setName("product");
        product.setAmount(5L);
        when(productService.findByName("product"))
                .thenReturn(Optional.of(product));

        Assertions.assertThat(
                testSubject.editProductInOrder(new EditeProductInOrderRequestDto(
                        1L,
                        "product",
                        8L,
                        ProductInOrderStatus.ACTIVE
                ))
        ).isEqualTo(new ProductInOrderDto(List.of("Not enough product - product in the amount of 8")));

        verify(productInOrderService).findById(1L);
        verify(productService).findByName("product");
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testEditProductInOrderWhenCancelled() {
        final ProductInOrder productInOrder = new ProductInOrder();
        final Order order = new Order();
        order.setId(1L);
        productInOrder.setOrder(order);
        final Product oldProduct = new Product();
        oldProduct.setId(2L);
        oldProduct.setName("old");
        oldProduct.setAmount(10L);
        productInOrder.setProduct(oldProduct);
        productInOrder.setAmount(5L);
        productInOrder.setProductInOrderStatus(ProductInOrderStatus.ACTIVE);
        when(productInOrderService.findById(1L))
                .thenReturn(Optional.of(productInOrder));

        final Product product = new Product();
        product.setId(1L);
        product.setAmount(10L);
        when(productService.findByName("product"))
                .thenReturn(Optional.of(product));

        when(productInOrderService.update(new UpdateProductInOrderParams(
                1L,
                1L,
                1L,
                8L,
                ProductInOrderStatus.CANCELLED
        ))).thenReturn(new ProductInOrder());

        when(productInOrderMapper.map(new ProductInOrder()))
                .thenReturn(new ProductInOrderDto(List.of("test")));

        Assertions.assertThat(
                testSubject.editProductInOrder(new EditeProductInOrderRequestDto(
                        1L,
                        "product",
                        8L,
                        ProductInOrderStatus.CANCELLED
                ))
        ).isEqualTo(new ProductInOrderDto(List.of("test")));

        verify(productInOrderService).findById(1L);
        verify(productService).findByName("product");
        verify(productInOrderService).update(new UpdateProductInOrderParams(
                1L,
                1L,
                1L,
                8L,
                ProductInOrderStatus.CANCELLED
        ));
        verify(productService).update(new UpdateProductParams(
                2L,
                "old",
                15L
        ));
        verify(productInOrderMapper).map(new ProductInOrder());
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testEditProductInOrderWhenStillActive() {
        final ProductInOrder productInOrder = new ProductInOrder();
        final Order order = new Order();
        order.setId(1L);
        productInOrder.setOrder(order);
        final Product oldProduct = new Product();
        oldProduct.setId(2L);
        oldProduct.setName("old");
        oldProduct.setAmount(10L);
        productInOrder.setProduct(oldProduct);
        productInOrder.setAmount(5L);
        productInOrder.setProductInOrderStatus(ProductInOrderStatus.ACTIVE);
        when(productInOrderService.findById(1L))
                .thenReturn(Optional.of(productInOrder));

        final Product product = new Product();
        product.setId(1L);
        product.setName("new");
        product.setAmount(10L);
        when(productService.findByName("product"))
                .thenReturn(Optional.of(product));

        when(productInOrderService.update(new UpdateProductInOrderParams(
                1L,
                1L,
                1L,
                8L,
                ProductInOrderStatus.ACTIVE
        ))).thenReturn(new ProductInOrder());

        when(productInOrderMapper.map(new ProductInOrder()))
                .thenReturn(new ProductInOrderDto(List.of("test")));

        Assertions.assertThat(
                testSubject.editProductInOrder(new EditeProductInOrderRequestDto(
                        1L,
                        "product",
                        8L,
                        ProductInOrderStatus.ACTIVE
                ))
        ).isEqualTo(new ProductInOrderDto(List.of("test")));

        verify(productInOrderService).findById(1L);
        verify(productService).findByName("product");
        verify(productInOrderService).update(new UpdateProductInOrderParams(
                1L,
                1L,
                1L,
                8L,
                ProductInOrderStatus.ACTIVE
        ));
        verify(productService).update(new UpdateProductParams(
                2L,
                "old",
                15L
        ));
        verify(productService).update(new UpdateProductParams(
                1L,
                "new",
                2L
        ));
        verify(productInOrderMapper).map(new ProductInOrder());
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testEditeOrderWhenOrderNotFound() {
        when(orderService.findById(1L))
                .thenReturn(Optional.empty());

        Assertions.assertThat(
                testSubject.editeOrder(new EditeOrderRequestDto(
                        1L,
                        OrderStatus.OPEN
                ))
        ).isEqualTo(new OrderDto(List.of("Not found order with id - 1")));

        verify(orderService).findById(1L);
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testEditeOrderWhenStillOpen() {
        final Order order = new Order();
        final Table table = new Table();
        table.setId(1L);
        order.setTable(table);
        when(orderService.findById(1L))
                .thenReturn(Optional.of(order));

        when(orderService.update(new UpdateOrderParams(
                1L,
                1L,
                OrderStatus.OPEN
        ))).thenReturn(new Order());

        when(orderMapper.map(new Order()))
                .thenReturn(new OrderDto(List.of("test")));

        Assertions.assertThat(
                testSubject.editeOrder(new EditeOrderRequestDto(
                        1L,
                        OrderStatus.OPEN
                ))
        ).isEqualTo(new OrderDto(List.of("test")));

        verify(orderService).findById(1L);
        verify(orderService).update(new UpdateOrderParams(
                1L,
                1L,
                OrderStatus.OPEN
        ));
        verify(orderMapper).map(new Order());
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testEditeOrderWhenClosed() {
        final Order order = new Order();
        order.setId(1L);
        final Table table = new Table();
        table.setId(1L);
        order.setTable(table);
        when(orderService.findById(1L))
                .thenReturn(Optional.of(order));

        when(orderService.update(new UpdateOrderParams(
                1L,
                1L,
                OrderStatus.CLOSED
        ))).thenReturn(new Order());

        final ProductInOrder productInOrder = new ProductInOrder();
        productInOrder.setId(1L);
        productInOrder.setOrder(order);
        final Product product = new Product();
        product.setId(1L);
        productInOrder.setProduct(product);
        productInOrder.setAmount(10L);
        when(productInOrderService.findByOrderId(1L))
                .thenReturn(List.of(productInOrder));

        when(orderMapper.map(new Order()))
                .thenReturn(new OrderDto(List.of("test")));

        Assertions.assertThat(
                testSubject.editeOrder(new EditeOrderRequestDto(
                        1L,
                        OrderStatus.CLOSED
                ))
        ).isEqualTo(new OrderDto(List.of("test")));

        verify(orderService).findById(1L);
        verify(orderService).update(new UpdateOrderParams(
                1L,
                1L,
                OrderStatus.CLOSED
        ));
        verify(productInOrderService).findByOrderId(1L);
        verify(productInOrderService).update(new UpdateProductInOrderParams(
                1L,
                1L,
                1L,
                10L,
                ProductInOrderStatus.CLOSED
        ));
        verify(orderMapper).map(new Order());
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }

    @Test
    public void testEditeOrderWhenCancelled() {
        final Order order = new Order();
        order.setId(1L);
        final Table table = new Table();
        table.setId(1L);
        order.setTable(table);
        when(orderService.findById(1L))
                .thenReturn(Optional.of(order));

        when(orderService.update(new UpdateOrderParams(
                1L,
                1L,
                OrderStatus.CANCELLED
        ))).thenReturn(new Order());

        final ProductInOrder productInOrder = new ProductInOrder();
        productInOrder.setId(1L);
        productInOrder.setOrder(order);
        final Product product = new Product();
        product.setId(1L);
        product.setName("product");
        product.setAmount(5L);
        productInOrder.setProduct(product);
        productInOrder.setAmount(10L);
        when(productInOrderService.findByOrderId(1L))
                .thenReturn(List.of(productInOrder));

        when(orderMapper.map(new Order()))
                .thenReturn(new OrderDto(List.of("test")));

        Assertions.assertThat(
                testSubject.editeOrder(new EditeOrderRequestDto(
                        1L,
                        OrderStatus.CANCELLED
                ))
        ).isEqualTo(new OrderDto(List.of("test")));

        verify(orderService).findById(1L);
        verify(orderService).update(new UpdateOrderParams(
                1L,
                1L,
                OrderStatus.CANCELLED
        ));
        verify(productInOrderService).findByOrderId(1L);
        verify(productInOrderService).update(new UpdateProductInOrderParams(
                1L,
                1L,
                1L,
                10L,
                ProductInOrderStatus.CANCELLED
        ));
        verify(productService).update(new UpdateProductParams(
                1L,
                "product",
                15L
        ));
        verify(orderMapper).map(new Order());
        verifyNoMoreInteractions(
                tableService,
                orderService,
                productInOrderService,
                productService,
                userService,
                tableMapper,
                orderMapper,
                productInOrderMapper
        );
    }
}